target/classes/scripts/GUARDRAIL-3/requirements.txt
Flask
pandas
joblib
scikit-learn
faiss-cpu
Werkzeug
sentence_transformers
openpyxl
